Content/PythonScripting/requirement.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
pillow==10.4.0
requests==2.32.3
urllib3==2.2.2
python-dotenv==1.0.1
pydub==0.25.1